Basic heuristic

+++ src/ai/MCTSMapEntry.java
@@ -8,18 +8,18 @@ import utils.Pair;
 
 public class MCTSMapEntry {
 	public int N;
-	public float E;
+	public double E;
 	public Set<List<Pair<Integer, Integer>>> V;
-	public Map<List<Pair<Integer, Integer>>, Float> P;
+	public Map<List<Pair<Integer, Integer>>, Double> P;
 	
 	public MCTSMapEntry() {
 		this.N = Integer.MIN_VALUE;
-		this.E = -Float.MAX_VALUE;
+		this.E = -Double.MAX_VALUE;
 		this.V = null;
 		this.P = null;
 	}
 	
-	public MCTSMapEntry(int N, float E, Set<List<Pair<Integer, Integer>>> V, Map<List<Pair<Integer, Integer>>, Float> P) {
+	public MCTSMapEntry(int N, float E, Set<List<Pair<Integer, Integer>>> V, Map<List<Pair<Integer, Integer>>, Double> P) {
 		this.N = N;
 		this.E = E;
 		this.V = V;
